requirements.txt
flask
Flask-WTF
flask-sqlalchemy
flask-socketio
gevent-websocket

